// src/serverApi/v3/models/room-member-response.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Schulcloud-Verbund-Software Server API
 * This is v3 of Schulcloud-Verbund-Software Server. Checkout /docs for v1.
 *
 * The version of the OpenAPI document: 3.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


import { RoleName } from './role-name';
import { RoomMemberResponseAllowedOperations } from './room-member-response-allowed-operations';

/**
 * 
 * @export
 * @interface RoomMemberResponse
 */
export interface RoomMemberResponse {
    /**
     * 
     * @type {string}
     * @memberof RoomMemberResponse
     */
    firstName: string;
    /**
     * 
     * @type {string}
     * @memberof RoomMemberResponse
     */
    lastName: string;
    /**
     * 
     * @type {RoleName}
     * @memberof RoomMemberResponse
     */
    roomRoleName: RoleName;
    /**
     * 
     * @type {Array<RoleName>}
     * @memberof RoomMemberResponse
     */
    schoolRoleNames: Array<RoleName>;
    /**
     * 
     * @type {string}
     * @memberof RoomMemberResponse
     */
    schoolId: string;
    /**
     * 
     * @type {string}
     * @memberof RoomMemberResponse
     */
    schoolName: string;
    /**
     * 
     * @type {string}
     * @memberof RoomMemberResponse
     */
    userId: string;
    /**
     * 
     * @type {RoomMemberResponseAllowedOperations}
     * @memberof RoomMemberResponse
     */
    allowedOperations?: RoomMemberResponseAllowedOperations;
}